msm8998: doze: use ExecutorService for listener registration
Registering a sensor seems to be an expensive operation,
and we do it on each screen-on event, so moving it to
an asynchronous task looks like a good idea anyway.

By moving all non-essential binder calls of the main thread or to the
next frame, we bring this down to 5ms, such that window animation
and Keyguard animation starts about at the same time.
https://developer.android.com/reference/java/util/concurrent/ExecutorService

Change-Id: I4f37bb9a7dc9d7775d587d4ebd4b6619f3b77e81